Monitor display/requirements.txt
requests
tk
Pillow
screeninfo
psutil
paramiko
